requirements.txt
fastapi==0.115.12
fastmcp==2.3.0
langfuse==2.60.7
mcp==1.11.0
pydantic==2.11.7
strands-agents==1.0.0
strands-agents-tools==0.1.8
uvicorn==0.35.0
openmeteo-requests==1.5.0
numpy==2.3.1
pandas==2.3.1
requests-cache==1.2.1
retry-requests==2.0.0